# BackendApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Delivery Backend API",
    "description": "AI-Optimized Smart Delivery Routing System — Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/MenuCategories/shop/{shopId}": {
      "get": {
        "tags": [
          "MenuCategories"
        ],
        "summary": "ดึงรายการหมวดหมู่สินค้าทั้งหมดของร้านค้าเดี่ยว (เรียงตามลำดับแสดงผล)",
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuCategoryDtoListApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/MenuCategories/{id}": {
      "put": {
        "tags": [
          "MenuCategories"
        ],
        "summary": "อัปเดตข้อมูลหมวดหมู่สินค้า",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuCategoryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuCategoryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuCategoryDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "MenuCategories"
        ],
        "summary": "ดึงข้อมูลตาม Primary Key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Primary Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuCategoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MenuCategories"
        ],
        "summary": "ลบข้อมูล (Soft Delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Primary Key ของข้อมูลที่ต้องการลบ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Riders/{id}": {
      "get": {
        "tags": [
          "Riders"
        ],
        "summary": "ดึงข้อมูลไรเดอร์เดี่ยว (รองรับทั้ง UUID และ Tracking Code)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiderDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Riders"
        ],
        "summary": "อัปเดตข้อมูลไรเดอร์ — อัปเดตเฉพาะฟิลด์ที่แก้ไขได้ เช่น ชื่อ\r\nป้องกันปัญหาในการเขียนทับ Read-Only/Init-Only properties เช่น RefNumber",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiderDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RiderDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RiderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiderDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Riders"
        ],
        "summary": "ลบข้อมูล (Soft Delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Primary Key ของข้อมูลที่ต้องการลบ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errorDetail": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API Response wrapper — ห่อทุก Response ให้อยู่ในรูปแบบเดียวกัน"
      },
...
      "MenuCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "trackingCode": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "shopId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO สำหรับแสดงข้อมูลหมวดหมู่เมนู"
      },
      "MenuCategoryDtoListApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errorDetail": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MenuCategoryDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API Response wrapper with payload"
      },
...
      "RiderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "รหัสไรเดอร์ (GUID)",
            "nullable": true
          },
          "trackingCode": {
            "type": "string",
            "description": "โค้ดไรเดอร์สำหรับการอ้างอิงอักษร",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "ชื่อไรเดอร์",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "สถานะไรเดอร์: AVAILABLE, DELIVERING, OFFLINE",
            "nullable": true
          },
          "lat": {
            "type": "number",
            "description": "ละติจูดปัจจุบัน",
            "format": "double",
            "nullable": true
          },
          "lng": {
            "type": "number",
            "description": "ลองจิจูดปัจจุบัน",
            "format": "double",
            "nullable": true
          },
          "lastUpdated": {
            "type": "string",
            "description": "เวลาที่อัปเดตตำแหน่งล่าสุด",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "DTO สำหรับส่งข้อมูล Rider ไปยัง Frontend"
      },
...
      "UpdateMenuCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "maximum": 1000,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO สำหรับอัปเดตหมวดหมู่เมนู"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter a valid JWT access token.",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}